pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["tiway <stilon670@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
django-rest-framework = "^0.1.0"
black = {extras = ["d"], version = "^24.10.0"}
pillow = "^10.4.0"
django-extensions = "^3.2.3"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
